delete_empty.sh: Don't clean out root filesystem if shared folder does not exist
As OpenWrt still needs a more verbose script then a simple find command,
we need to stick at this. The script now exits if the target directory does
not exist.

Also we try to reduce the amount of files checked in the while clause with
reducing the find output to regular files only.

Closed: #133