lighttpd upload: align Upload sizes
With the introduction of PHP and as a preparation for the
multipart uploader, we moved the tmp space of lighttpd
to the ramdisk on the tiny routers.
We now need to adjust the upload size in kareha to that limit.

With the same commit, I raize the MB limit of lighttpd to 5MB,
which is a good compromise.