Droopy: Make chmod optional.
Droopy always does the chmod operation, if the option is set.
On partitions with FAT it is not possible to execute this operation.
Droopy may ran into issue in that situation.

With that fix, the option won't be set on droopy start to prevent
any chmod operation on FAT partitions.

Fixes #154